requirements (1).txt
fastapi
uvicorn[standard]
openai
faiss-cpu
python-dotenv
pydantic
fhirclient
pinecone-client
